// block-performance-1.0.0/src/library/packager/npm-package-analyze/type.ts
// auto generated by dev/system

import { Logger } from "../../../myutils/logger.js";

export interface Input {
    input_dir: string;
}

export interface OutputOk {
    npm_package_list: { name: string }[];
}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}

export function check_input<R>(plog: Logger, v: any, cb: { ok: () => R; fail: (err: Error) => R }): R {
    const log = plog.sub("check_input");
    log.variable("v", v);
    try {
        log.println("v must be object");
        if (typeof v !== "object" || v === null) {
            throw new Error("v is not object");
        }

        Object.keys(v).forEach((field) => {
            if (field === "input_dir") return;
            throw new Error("v contains unknown field: " + field);
        });

        log.println("v.input_dir must be string");
        if (typeof v.input_dir !== "string") {
            throw new Error("v.input_dir is not string");
        }
    } catch (err) {
        log.error(err);
        return cb.fail(err);
    }
    // nothing wrong
    log.ok();
    return cb.ok();
}

// JSON stringify value before copy to handle some specific problem
// eg. Date Object probelm
export function copy_output_ok(v: any): OutputOk {
    if (v !== undefined) {
        const v_cloned = JSON.parse(JSON.stringify(v));
        return _copy_output_ok(v_cloned);
    } else {
        return _copy_output_ok(v);
    }
}

export function _copy_output_ok(v: any): OutputOk {
    if (typeof v === "object" && v !== null) {
        const obj = {
            npm_package_list: copy_npm_package_list(v.npm_package_list)
        };
        return obj;
    } else {
        return { npm_package_list: [] };
    }

    function copy_npm_package_list(v: any): { name: string }[] {
        return Array.isArray(v) ? v.map(copy_item) : [];

        function copy_item(v: any): { name: string } {
            if (typeof v === "object" && v !== null) {
                const obj = {
                    name: copy_name(v.name)
                };
                return obj;
            } else {
                return { name: "" };
            }

            function copy_name(v: any): string {
                return typeof v === "string" ? v : "";
            }
        }
    }
}

// backward compatible purpose
export type Output = OutputOk;
export const copy_output = copy_output_ok;